fix: fine alignment details
