+++ packages/rainbowkit/src/components/Dialog/DialogContent.tsx
@@ -1,4 +1,5 @@
 import React, { ReactNode } from 'react';
+import { isMobile } from '../../utils/isMobile';
 import { Box, BoxProps } from '../Box/Box';
 import * as styles from './DialogContent.css';
 
@@ -7,6 +8,7 @@ interface DialogContentProps {
   bottomSheetOnMobile?: boolean;
   padding?: BoxProps['padding'];
   marginTop?: BoxProps['marginTop'];
+  wide?: boolean;
 }
 
 export function DialogContent({
@@ -14,12 +16,14 @@ export function DialogContent({
   children,
   marginTop,
   padding = '16',
+  wide = false,
 }: DialogContentProps) {
+  const mobile = isMobile();
   return (
     <Box marginTop={marginTop}>
       <Box
         className={[
-          styles.dialogContent,
+          mobile || wide ? styles.dialogContentWide : styles.dialogContent,
           bottomSheetOnMobile ? styles.bottomSheetOverrides : null,
         ].join(' ')}
       >
